+++ examples/1App_ACCERT_PWR12BE/watts_exec.py
@@ -16,7 +16,7 @@ params['cost_217'] = 29000000
 input_name = "ACCERT_input.tmpl"
 
 
-accert_plugin = watts.PluginACCERT(f'{input_name}')
+accert_plugin = watts.PluginACCERT(input_name)
 accert_result = accert_plugin(params)
 
 print(accert_result.inputs)